Allow using `rfl::AddTagsToVariants` with rfl::Generic and same-name structs
The first one was the main reason for this change; the second one was a result
of applying the change. Added tests for both.

The issue was that both `std::vector<Generic>` and `rfl::Object<Generic>` were
getting the type name 'Generic', which caused a compiler error
due to duplicated field names.

This will slightly increase the size of serialized formats when using the
AddTagsToVariants, but it seems like a worthwhile change.